comprocess/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
colorama==0.4.6
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
distro==1.9.0
Django==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.7
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
joblib==1.4.2
numpy==1.26.4
openai==2.8.1
pandas==2.2.3
pillow==11.1.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
scikit-learn==1.5.2
sniffio==1.3.1
sqlparse==0.5.3
tensorflow==2.18.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.8.2
pillow==11.1.0
